Add advanced app search UI.
