+++ src/pages/EntityDetails/Model/ApplicationsTab/LocalAppsTable/AppSearchBox/AppSearchBox.tsx
@@ -1,12 +1,14 @@
-import { SearchBox } from "@canonical/react-components";
+import { SearchBox, Icon, Switch, Tooltip } from "@canonical/react-components";
 import { useState } from "react";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 
 import { useQueryParams } from "hooks/useQueryParams";
+import { getConfig } from "store/general/selectors";
 import { actions as jujuActions } from "store/juju";
 
 const AppSearchBox = () => {
   const dispatch = useDispatch();
+  const isJuju = useSelector(getConfig)?.isJuju;
   const [query, setQuery] = useQueryParams<{
     filterQuery: string;
     filterType: string | null;
@@ -29,15 +31,19 @@ const AppSearchBox = () => {
 
   return (
     <form
-      className="u-flex-grow"
+      className="u-flex u-flex-grow"
       onSubmit={(event) => {
         event.preventDefault();
         updateQuery(filterQuery);
       }}
     >
       <SearchBox
-        className="u-no-margin"
-        placeholder="Filter applications"
+        className="u-no-margin u-flex-grow"
+        placeholder={
+          query.filterType === "jq"
+            ? "Filter applications using JQ"
+            : "Filter applications"
+        }
         onSearch={() => updateQuery(filterQuery)}
         onClear={() => updateQuery("")}
         onChange={setFilterQuery}
@@ -45,6 +51,29 @@ const AppSearchBox = () => {
         data-testid="filter-applications"
         value={filterQuery}
       />
+      {!isJuju ? (
+        <div className="u-sph--large u-no-margin--right">
+          <Switch
+            checked={query.filterType === "jq"}
+            label={
+              <>
+                JQ{" "}
+                <Tooltip
+                  position="btm-left"
+                  //   TODO: add a link to the docs
+                  message="Filter applications in this model using JQ. Find out more."
+                >
+                  <Icon name="help" />
+                </Tooltip>
+              </>
+            }
+            onChange={(event) => {
+              const target = event.target as HTMLInputElement;
+              setQuery({ filterType: target.checked ? "jq" : null });
+            }}
+          />
+        </div>
+      ) : null}
     </form>
   );
 };
