+++ src/pages/EntityDetails/Model/ApplicationsTab/LocalAppsTable/LocalAppsTable.test.tsx
@@ -265,84 +265,4 @@ describe("LocalAppsTable", () => {
     );
     expect(screen.getByTestId("select-all-apps")).not.toBeChecked();
   });
-
-  it("doesn't show the run action button when there is no search", () => {
-    renderComponent(
-      <LocalAppsTable
-        applications={state.juju.modelWatcherData?.test123.applications}
-      />,
-      { path, url, state }
-    );
-    expect(
-      screen.queryByRole("button", { name: /run action/i })
-    ).not.toBeInTheDocument();
-  });
-
-  it("shows the run action button when there is a search", () => {
-    renderComponent(
-      <LocalAppsTable
-        applications={state.juju.modelWatcherData?.test123.applications}
-      />,
-      {
-        path,
-        url: searchURL,
-        state,
-      }
-    );
-    expect(
-      screen.queryByRole("button", { name: /run action/i })
-    ).toBeInTheDocument();
-  });
-
-  it("disables the run action button when there are no applications selected", () => {
-    renderComponent(
-      <LocalAppsTable
-        applications={state.juju.modelWatcherData?.test123.applications}
-      />,
-      {
-        path,
-        url: searchURL,
-        state,
-      }
-    );
-    expect(
-      screen.queryByRole("button", { name: /run action/i })
-    ).toBeDisabled();
-  });
-
-  it("enables the run action button when there is at least one application selected", () => {
-    state.juju.selectedApplications = [charmApplicationFactory.build()];
-    renderComponent(
-      <LocalAppsTable
-        applications={state.juju.modelWatcherData?.test123.applications}
-      />,
-      {
-        path,
-        url: searchURL,
-        state,
-      }
-    );
-    expect(
-      screen.queryByRole("button", { name: /run action/i })
-    ).not.toBeDisabled();
-  });
-
-  it("opens the choose-charm panel when clicking the run action button", async () => {
-    state.juju.selectedApplications = [charmApplicationFactory.build()];
-    renderComponent(
-      <LocalAppsTable
-        applications={state.juju.modelWatcherData?.test123.applications}
-      />,
-      {
-        path,
-        url: searchURL,
-        state,
-      }
-    );
-    expect(window.location.search).toEqual("?filterQuery=db");
-    await userEvent.click(screen.getByRole("button", { name: /run action/i }));
-    expect(window.location.search).toEqual(
-      "?filterQuery=db&panel=select-charms-and-actions"
-    );
-  });
 });
