+++ src/pages/EntityDetails/Model/ApplicationsTab/LocalAppsTable/SearchResultsActionsRow/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SearchResultsActionsRow";
