+++ src/scss/_utils.scss
@@ -60,3 +60,8 @@
   padding-left: $sph--large;
   padding-right: $sph--large;
 }
+
+.u-sph--large {
+  margin-left: $sph--large;
+  margin-right: $sph--large;
+}
